Move data fetching into client component
